Update migration feature
